Add seg::{insert,display,table}
